+++ src/patch/core/Global.zig
@@ -14,10 +14,12 @@ const xinput = @import("../util/xinput.zig");
 const dbg = @import("../util/debug.zig");
 const msg = @import("../util/message.zig");
 const mem = @import("../util/memory.zig");
-const r = @import("../util/racer.zig");
-const rf = r.functions;
-const rc = r.constants;
-const rt = r.text;
+
+const rti = @import("racer").Time;
+const rg = @import("racer").Global;
+const rrd = @import("racer").RaceData;
+const re = @import("racer").Entity;
+const rt = @import("racer").Text;
 const rto = rt.TextStyleOpts;
 
 const w32 = @import("zigwin32");
@@ -109,11 +111,10 @@ pub const GlobalState = extern struct {
 
     fn player_reset(self: *GlobalState) void {
         const p = &self.player;
-        const u: [14]u8 = mem.deref_read(&.{ 0x4D78A4, 0x0C, 0x41 }, [14]u8);
-        p.upgrades_lv = u[0..7].*;
-        p.upgrades_hp = u[7..14].*;
+        p.upgrades_lv = rrd.PLAYER.*.pFile.upgrade_lv; // TODO: remove from gs, now that it's easy?
+        p.upgrades_hp = rrd.PLAYER.*.pFile.upgrade_hp; // TODO: remove from gs, now that it's easy?
         p.upgrades = for (0..7) |i| {
-            if (u[i] > 0 and u[7 + i] > 0) break true;
+            if (p.upgrades_lv[i] > 0 and p.upgrades_hp[i] > 0) break true;
         } else false;
 
         p.flags1 = 0;
@@ -123,16 +124,16 @@ pub const GlobalState = extern struct {
         p.dead = .Off;
         p.deaths = 0;
 
-        p.heat_rate = r.ReadPlayerValue(0x8C, f32);
-        p.cool_rate = r.ReadPlayerValue(0x90, f32);
+        p.heat_rate = re.Test.PLAYER.*.stats.HeatRate; // TODO: remove from gs, now that it's easy?
+        p.cool_rate = re.Test.PLAYER.*.stats.CoolRate; // TODO: remove from gs, now that it's easy?
         p.heat = 0;
     }
 
     fn player_update(self: *GlobalState) void {
         const p = &self.player;
-        p.flags1 = r.ReadPlayerValue(0x60, u32);
-        p.heat = r.ReadPlayerValue(0x218, f32);
-        const engine: [6]u32 = r.ReadPlayerValue(0x2A0, [6]u32);
+        p.flags1 = re.Test.PLAYER.*.flags1; // TODO: remove from gs, now that it's easy?
+        p.heat = re.Test.PLAYER.*.temperature; // TODO: remove from gs, now that it's easy?
+        const engine = re.Test.PLAYER.*.engineStatus; // TODO: remove from gs, now that it's easy?
 
         p.boosting.update((p.flags1 & (1 << 23)) > 0);
         p.underheating.update(p.heat >= 100);
@@ -195,8 +196,9 @@ pub fn init() bool {
     if (kb_shift_dn)
         return false;
 
-    GLOBAL_STATE.hwnd = mem.read(rc.ADDR_HWND, win.HWND);
-    GLOBAL_STATE.hinstance = mem.read(rc.ADDR_HINSTANCE, win.HINSTANCE);
+    // TODO: remove? probably don't need these anymore lol
+    GLOBAL_STATE.hwnd = rg.HWND.*;
+    GLOBAL_STATE.hinstance = rg.HINSTANCE.*;
 
     return true;
 }
@@ -208,20 +210,19 @@ pub fn OnInitLate(gs: *GlobalState, _: *GlobalFunction) callconv(.C) void {
 }
 
 pub fn EngineUpdateStage14A(gs: *GlobalState, _: *GlobalFunction) callconv(.C) void {
-    const player_ready: bool = mem.read(rc.RACE_DATA_PLAYER_RACE_DATA_PTR_ADDR, u32) != 0 and
-        r.ReadRaceDataValue(0x84, u32) != 0;
+    const player_ready: bool = rrd.PLAYER_PTR.* != 0 and rrd.PLAYER.*.pTestEntity != 0;
     gs.in_race.update(player_ready);
 
     gs.race_state_prev = gs.race_state;
     gs.race_state = blk: {
         if (!gs.in_race.on()) break :blk .None;
-        if (mem.read(rc.ADDR_IN_RACE, u8) == 0) break :blk .PreRace;
+        if (rg.IN_RACE.* == 0) break :blk .PreRace;
         // TODO: figure out how the engine knows to set these and use those instead
-        const flags1 = r.ReadPlayerValue(0x60, u32);
+        const flags1 = re.Test.PLAYER.*.flags1;
         const countdown: bool = flags1 & (1 << 0) != 0;
         if (countdown) break :blk .Countdown;
         const postrace: bool = flags1 & (1 << 5) == 0;
-        const show_stats: bool = r.ReadEntityValue(.Jdge, 0, 0x08, u32) & 0x0F == 2;
+        const show_stats: bool = re.Manager.entity(.Jdge, 0).flags & 0x0F == 2;
         if (postrace and show_stats) break :blk .PostRace;
         if (postrace) break :blk .PostRaceExiting;
         break :blk .Racing;
@@ -233,12 +234,12 @@ pub fn EngineUpdateStage14A(gs: *GlobalState, _: *GlobalFunction) callconv(.C) v
 }
 
 pub fn TimerUpdateA(gs: *GlobalState, _: *GlobalFunction) callconv(.C) void {
-    gs.dt_f = mem.read(rc.ADDR_TIME_FRAMETIME, f32);
-    gs.fps = mem.read(rc.ADDR_TIME_FPS, f32);
+    gs.dt_f = rti.FRAMETIME.*;
+    gs.fps = rti.FPS.*;
     const fps_res: f32 = 1 / gs.dt_f * 2;
     gs.fps_avg = (gs.fps_avg * (fps_res - 1) + (1 / gs.dt_f)) / fps_res;
-    gs.timestamp = mem.read(rc.ADDR_TIME_TIMESTAMP, u32);
-    gs.framecount = mem.read(rc.ADDR_TIME_FRAMECOUNT, u32);
+    gs.timestamp = rti.TIMESTAMP.*;
+    gs.framecount = rti.FRAMECOUNT.*;
 }
 
 pub fn MenuTitleScreenB(_: *GlobalState, _: *GlobalFunction) callconv(.C) void {