+++ src/patch/util/scroll_control.zig
@@ -8,9 +8,9 @@ const win32kb = win32.ui.input.keyboard_and_mouse;
 const st = @import("active_state.zig");
 const mem = @import("memory.zig");
 const input = @import("../core/input.zig");
-const r = @import("racer.zig");
-const rc = r.constants;
-const rf = r.functions;
+
+const r = @import("racer");
+const rt = r.Time;
 
 pub const InputGetFnType = *const fn (st.ActiveState) callconv(.C) bool;
 
@@ -23,7 +23,7 @@ pub const ScrollControl = extern struct {
     input_inc: InputGetFnType,
 
     pub fn UpdateEx(self: *ScrollControl, val: i32, max: i32, wrap: bool) i32 {
-        const dt = mem.read(rc.ADDR_TIME_FRAMETIME, f32);
+        const dt = rt.FRAMETIME.*;
 
         var inc: f32 = 0;
         if (self.input_dec(.On)) self.scroll -= dt;