+++ src/racer/Video.zig
@@ -1,0 +1,13 @@
+// GAME FUNCTIONS
+
+// NOTE: strings found at 0x4B7A44 in game memory
+/// @filename   a .znm file located in /data/anims
+pub const swrVideo_PlayVideoFile: *fn (filename: [*:0]const u8) callconv(.C) void = @ptrFromInt(0x4252A0);
+
+// GAME CONSTANTS
+
+// ...
+
+// HELPERS
+
+// ...