+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Migrated `patch/util/racer*` to `racer` module
+
 ## [0.1.4] - 2024-05-09
 
 ### Added
@@ -44,7 +50,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	- Can now hold button to open instead of double-tap
 	- Scroll through FPS presets with hotkey
 	- Scroll through tracks by planet with hotkey
-
 
 ### Changed